Added module support.
Modules must be instantiable and not static.  Modules must also have a
public variable $name to be registered.  Classnames will be dynamically
generated.  Please use "@@CLASSNAME@@" in place of the actual class
name.  Modules also need to have a function isInstantiated() that
returns true.  This function should be used as an initialization
function.